chore: add types to all control panels
