+++ plugins/legacy-plugin-chart-calendar/src/controlPanel.ts
@@ -16,15 +16,16 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+import { t, legacyValidateInteger } from '@superset-ui/core';
 import {
-  t,
-  legacyValidateInteger,
-  D3_TIME_FORMAT_OPTIONS,
+  ControlPanelConfig,
   D3_FORMAT_DOCS,
-} from '@superset-ui/core';
-import { formatSelectOptions, sections } from '@superset-ui/chart-controls';
+  D3_TIME_FORMAT_OPTIONS,
+  formatSelectOptions,
+  sections,
+} from '@superset-ui/chart-controls';
 
-export default {
+const config: ControlPanelConfig = {
   controlPanelSections: [
     sections.legacyRegularTime,
     {
@@ -177,3 +178,5 @@ export default {
     },
   },
 };
+
+export default config;
