+++ plugins/legacy-plugin-chart-chord/src/controlPanel.ts
@@ -17,9 +17,9 @@
  * under the License.
  */
 import { t, validateNonEmpty } from '@superset-ui/core';
-import { sections } from '@superset-ui/chart-controls';
+import { ControlPanelConfig, sections } from '@superset-ui/chart-controls';
 
-export default {
+const config: ControlPanelConfig = {
   controlPanelSections: [
     sections.legacyRegularTime,
     {
@@ -55,3 +55,5 @@ export default {
     },
   },
 };
+
+export default config;
