+++ plugins/legacy-plugin-chart-force-directed/src/controlPanel.ts
