+++ plugins/legacy-plugin-chart-horizon/src/controlPanel.ts
@@ -17,9 +17,9 @@
  * under the License.
  */
 import { t } from '@superset-ui/core';
-import { formatSelectOptions, sections } from '@superset-ui/chart-controls';
+import { ControlPanelConfig, formatSelectOptions, sections } from '@superset-ui/chart-controls';
 
-export default {
+const config: ControlPanelConfig = {
   controlPanelSections: [
     sections.legacyRegularTime,
     {
@@ -92,3 +92,5 @@ export default {
     },
   ],
 };
+
+export default config;
