+++ plugins/legacy-plugin-chart-rose/src/controlPanel.tsx
@@ -19,14 +19,15 @@
 import React from 'react';
 import { t } from '@superset-ui/core';
 import {
+  ControlPanelConfig,
   formatSelectOptions,
-  D3_TIME_FORMAT_OPTIONS,
   D3_FORMAT_DOCS,
   D3_FORMAT_OPTIONS,
+  D3_TIME_FORMAT_OPTIONS,
   sections,
 } from '@superset-ui/chart-controls';
 
-export default {
+const config: ControlPanelConfig = {
   controlPanelSections: [
     sections.legacyTimeseriesTime,
     {
@@ -251,3 +252,5 @@ export default {
     },
   ],
 };
+
+export default config;
